% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{addClusters}
\alias{addClusters}
\title{Find Cluster using SNN graph clustering methods}
\usage{
addClusters(obj, minGroupSize = 50, dimsUse = seq_len(50),
  initialResolution = 0.8, nStart = 10, ...)
}
\arguments{
\item{obj}{Matrix  cell-by-feature matrix}

\item{minGroupSize}{the mininum cell required to be cluster,
200 is required according to 2019, NBT.
Set it to NULL if the peaks are called}

\item{initialResolution}{Value of the resolution parameter,
use a value above (below) 1.0 if you want to obtain a larger (smaller) number of communities}

\item{...}{Arguments passed to Seurat's FindNeighbors or FindClusters}

\item{dims.use}{Dimensions of reduction to use as input}

\item{n.start}{Number of random starts}
}
\value{
Seurat object
}
\description{
Find Cluster using SNN graph clustering methods
}
