% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedToBigWig.R
\name{bedToBigWig}
\alias{bedToBigWig}
\title{Convert BED to BigWig}
\usage{
bedToBigWig(bed, chromSizes = NULL, binSize = 100, minCell = 30,
  norm = "BPM", unionPeaks = NULL, scale = 1e+07, verbose = FALSE,
  ...)
}
\arguments{
\item{bed}{BED file path, which will be converted to BigWig}

\item{chromSizes}{GRanges store chromosome size}

\item{binSize}{bin size}

\item{minCell}{fiter cluster with cell less than minCell}

\item{norm}{BPM, readInPeak for samples of varying quality}

\item{unionPeaks}{unin peak sets in GRanges}

\item{scale}{scale factor}

\item{...}{other params}
}
\description{
Convert BED to BigWig
}
