% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{buildWindows}
\alias{buildWindows}
\title{Build window along the genome}
\usage{
buildWindows(seqname, size, windowSize = 5000)
}
\arguments{
\item{seqname}{sequence name}

\item{size}{sequence size}

\item{windowSize}{each window size}
}
\description{
Build window along the genome
}
