% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cicero.R
\name{calcConnections}
\alias{calcConnections}
\title{Computing grouped correlations}
\usage{
calcConnections(ciceroObj, flank = 250 * 10^3, ...)
}
\arguments{
\item{ciceroObj}{cicero cell_data_set object}
}
\description{
Computing grouped correlations
}
