% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countInsertions.R
\name{countInsertions}
\alias{countInsertions}
\title{Generate insertion count matrixy by feature}
\usage{
countInsertions(fragments, feature, by = "RG", ...)
}
\arguments{
\item{fragments}{GenomicRanges object}

\item{feature}{GenomicRanges object  to store feature information}

\item{by}{metadata in mcols}

\item{...}{Arguments passed to other methods}
}
\value{
An list with sparseMatrix, FRIP and total fragments
}
\description{
Generate insertion count matrixy by feature
}
