% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak.R
\name{createPeakSummarizedExperiment}
\alias{createPeakSummarizedExperiment}
\title{createPeakSummarizedExperiment}
\usage{
createPeakSummarizedExperiment(fragmentFiles, genome, groups = "scATAC",
  extend = 250, blacklist = NULL, nSummits = 2e+05,
  chromKeep = NULL, dirPeaks = "results/LSI-Cluster-Peaks",
  clusters = NULL, ...)
}
\arguments{
\item{fragmentFiles}{vector include}

\item{genome}{genomeSize GRanges, used in extendedPeakSet}

\item{extend}{extend, used in extendedPeakSet}

\item{blacklist}{blacklist, use in extendedPeakSet}

\item{nSummits}{nSummits, use in extendedPeakSet}

\item{chromKeep}{chromosome to keep}

\item{clusters}{clusters generated by Seurat SNN graph clustering}
}
\description{
createPeakSummarizedExperiment
}
