% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_peak_analysis.R
\name{createPseudoBulk}
\alias{createPseudoBulk}
\title{create pseudo bulk by cells in cluster}
\usage{
createPseudoBulk(mat, groups, labels, minCells = 100, maxCells = 500,
  minReps = 3, ceiling = 1, prior.count = 3, nSim = 1000,
  distMethod = "vars", seed = 1)
}
\arguments{
\item{mat}{cell-by-peak matrix}

\item{groups}{group, for example, cluster information}

\item{labels}{label}

\item{minCells}{min cell in pseduo bulk sample, default is 100}

\item{maxCells}{max cell in pseduo bulk sample, default is 500}

\item{minReps}{pseduo bulk sample number, default is 3}

\item{ceiling}{Setting ceiling of input matrix to,
default is 1 means binarize the matrix}

\item{prior.count}{add the count to the matrix}

\item{nSim}{number of simulations}

\item{distMethod}{distance calculate methods}

\item{seed}{random seet, default is 1}
}
\description{
create pseudo bulk by cells in cluster
}
