% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_peak_analysis.R
\name{createPseudoBulkSummarizedExperiment}
\alias{createPseudoBulkSummarizedExperiment}
\title{create Pseudo-Bulk SummarizedExperiment object}
\usage{
createPseudoBulkSummarizedExperiment(objPB, se)
}
\arguments{
\item{objPB}{object return from createPseduoBulk to DataFrame}

\item{se}{scATAC-Summarized-Experiment}
}
\description{
create Pseudo-Bulk SummarizedExperiment object
}
