% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak.R
\name{extendedPeakSet}
\alias{extendedPeakSet}
\title{Extend the peak}
\usage{
extendedPeakSet(df, genomeSize = NULL, extend = 250,
  blacklist = NULL, nSummits = 1e+05, ...)
}
\arguments{
\item{df}{data.frame object include peak information}

\item{genomeSize}{genome size GRange object}

\item{extend}{extend the summit}

\item{blacklist}{genome black list}

\item{nSummits}{number of summit}

\item{...}{Arguments passed to seurat RunLSI}
}
\value{
Seurat object
}
\description{
Extend the peak
}
