% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{filterFragments}
\alias{filterFragments}
\title{Filter Fragments by fragments and enrichment score}
\usage{
filterFragments(fragments, tssSingles, filterFrags = 1000,
  filterTSS = 8, prefix = NULL)
}
\arguments{
\item{fragments}{GRanges object}

\item{tssSingles}{tssSingle From getTssProfile}

\item{filterFrags}{threshold of fragments per cell}

\item{filterTSS}{threshold of enrichment score of TSS}

\item{prefix}{prefix add to barcode}
}
\description{
Filter Fragments by fragments and enrichment score
}
