% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{getTssProfile}
\alias{getTssProfile}
\title{TSS Profile}
\usage{
getTssProfile(fragments, txdb, batchSize = 1000)
}
\arguments{
\item{fragments}{a GRanges object}

\item{txdb}{txdb object}

\item{batchSize}{batch size}
}
\description{
TSS Profile
}
