% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cicero.R
\name{getTxDbGenes}
\alias{getTxDbGenes}
\title{get TxDb genes}
\usage{
getTxDbGenes(txdb = NULL, orgdb = NULL, gr = NULL,
  ignore.strand = TRUE)
}
\arguments{
\item{txdb}{txdb object}

\item{orgdb}{orgdb object}

\item{gr}{GRanges}

\item{ignore.strand}{ignore the strand}
}
\description{
get TxDb genes
}
