% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_peak_analysis.R
\name{groupMeans}
\alias{groupMeans}
\title{calcuate  means of each group}
\usage{
groupMeans(mat, groups = NULL, na.rm = TRUE, sparse = FALSE)
}
\arguments{
\item{mat}{cell-by-peak count matrix}

\item{groups}{group name vector}

\item{na.rm}{romove NA}

\item{sparse}{sparse matrix or matrix}
}
\description{
calcuate  means of each group
}
