% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{insertionProfileSingles}
\alias{insertionProfileSingles}
\title{Generate insertion profile from fragments}
\usage{
insertionProfileSingles(fragments, feature, by = "RG",
  getInsertions = TRUE, fix = "center", flank = 2000, norm = 100,
  smooth = 51, range = 100, batchSize = 100, ...)
}
\arguments{
\item{fragments}{GenomicRanges object}

\item{feature}{GenomicRanges object  to store feature information}

\item{by}{metadata in mcols}

\item{flank}{flank}

\item{smooth}{smooth}

\item{range}{range}

\item{batchSize}{Split the cell into N batch for performance}

\item{...}{Arguments passed to other methods}

\item{nrom}{nrom}
}
\value{
An list with sparseMatrix, FRIP and total fragments
}
\description{
Generate insertion profile from fragments
}
