% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{insertionProfileSingles_helper}
\alias{insertionProfileSingles_helper}
\title{insertionProfileSingles helper}
\usage{
insertionProfileSingles_helper(feature, fragments, by = "RG",
  getInsertions = TRUE, fix = "center", flank = 2000, norm = 100,
  smooth = 51, range = 100, batchSize = 100, ...)
}
\arguments{
\item{feature}{GenomicRanges object  to store feature information}

\item{fragments}{GenomicRanges object}

\item{by}{metadata in mcols}

\item{flank}{the set flank size of TSS to find the overlap of
insertion}

\item{smooth}{smooth}

\item{range}{range}

\item{batchSize}{batch size}

\item{...}{Arguments passed to other methods}

\item{nrom}{nrom}
}
\value{
An list with sparseMatrix, FRIP and total fragments
}
\description{
insertionProfileSingles helper
}
