% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionReduction.R
\name{lsi}
\alias{lsi}
\title{LSI reduction}
\usage{
lsi(mat, nComponents = 50, binarize = TRUE, nFeatures = 2000, ...)
}
\arguments{
\item{mat}{dgCMatrix object}

\item{nComponents}{Number of singular values to compute}

\item{binarize}{binarize the matrix}

\item{nFeatures}{Getting top N features for  SVD
if nFeatures < 1, Getting the features  expressing >= cell * nFeatures}

\item{...}{Arguments passed to seurat RunLSI}
}
\value{
SVD matrix
}
\description{
LSI reduction
}
