% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cicero.R
\name{makeCDS}
\alias{makeCDS}
\title{make cell_data_set object from SummarizedExperiment object}
\usage{
makeCDS(se, binarize = TRUE, ...)
}
\arguments{
\item{se}{SummarizedExperiment object.}

\item{binarize}{binary the matrix.}
}
\description{
make cell_data_set object from SummarizedExperiment object
}
