% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cicero.R
\name{makeCicero}
\alias{makeCicero}
\title{make Cicero cell_data_set object from cell_data_setobject}
\usage{
makeCicero(cds, k = 50, ...)
}
\arguments{
\item{k}{Number of cells to aggregate per bin.}

\item{se}{SummarizedExperiment object.}
}
\description{
make Cicero cell_data_set object from cell_data_setobject
}
