% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak.R
\name{mergePeak}
\alias{mergePeak}
\title{merge the peak from all sample}
\usage{
mergePeak(..., by = "scoreQuantile", decreasing = TRUE)
}
\arguments{
\item{...}{peak from different sample}

\item{by}{the overlapped peak order standard}

\item{decreasing}{logical. Should the sort order be increasing or decreasing?}

\item{verbose}{Be chatty and report processing?}
}
\description{
merge the peak from all sample
}
