% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak.R
\name{nonOverlappingGRanges}
\alias{nonOverlappingGRanges}
\title{only keep the most significant peak in overlap region}
\usage{
nonOverlappingGRanges(gr, by = "score", decreasing = TRUE,
  verbose = FALSE)
}
\arguments{
\item{gr}{GRanges object}

\item{by}{the overlapped peak order standard}

\item{decreasing}{logical. Should the sort order be increasing or decreasing?}

\item{verbose}{Be chatty and report processing?}
}
\description{
only keep the most significant peak in overlap region
}
