% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_GA_umap}
\alias{plot_GA_umap}
\title{plot gene activity score in UMAP}
\usage{
plot_GA_umap(cds, cicero_gene_activities, gene, dotSize = 1,
  scale = 1e+06)
}
\description{
plot gene activity score in UMAP
}
