% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotTssProfile}
\alias{plotTssProfile}
\title{Plot TSS Profile}
\usage{
plotTssProfile(tssSingles, filterFrags = 1000, filterTSS = 8,
  outFile = NULL)
}
\arguments{
\item{tssSingles}{object return from getTssProfile}

\item{filterFrags}{threshold of fragments of cell}

\item{filterTSS}{enrichment score of TSS}

\item{outFile}{graphics file path}
}
\description{
Plot TSS Profile
}
