% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{readFragments}
\alias{readFragments}
\title{Reading Fragment Files}
\usage{
readFragments(file = NULL, minFrags = 0, ...)
}
\arguments{
\item{file}{fragments file path}

\item{minFrags}{filter cell with few fragments}
}
\description{
Reading Fragment Files
}
