% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{readInPeakNorm}
\alias{readInPeakNorm}
\title{Calculate the factor of read in peak normalization}
\usage{
readInPeakNorm(gr, unionPeaks, scale = 1e+07, ...)
}
\arguments{
\item{unionPeaks}{union peak}

\item{fragmentFile}{file path of Tn5 offset-corrected(+4 for +strand and -5 for -strand) insertion sites}

\item{chromSize}{chrome size}
}
\description{
Calculate the factor of read in peak normalization
}
