% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak.R
\name{runMACS2}
\alias{runMACS2}
\title{Run MACS2 to call peak in pseduo-bulk}
\usage{
runMACS2(bed.path = NULL, peak.dir = NULL, macs2Path = "macs2",
  method = "q", cutoff = 0.05, shift = -75, extsize = 150,
  bedGraph = TRUE, genomeSize = 2.7e+09, addtion = NULL)
}
\arguments{
\item{bed.path}{input BED path}

\item{peak.dir}{output peak dir}

\item{macs2Path}{macs2 path}

\item{method}{q qvalue, p pvalue}

\item{cutoff}{cutoff of pvalue or qvalue}

\item{shift}{The arbitrary shift in bp}

\item{extsize}{The arbitrary extension size in bp}

\item{bedGraph}{output bedGraph}

\item{genomeSize}{Effective genome size}

\item{addtion}{additional options passed to MACS2}
}
\description{
Run MACS2 to call peak in pseduo-bulk
}
