% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{seuratLSI}
\alias{seuratLSI}
\title{LSI reduction analysis}
\usage{
seuratLSI(mat, nComponents = 50, binarize = TRUE, nFeatures = NULL,
  ...)
}
\arguments{
\item{mat}{GenomicRanges object  to store feature information}

\item{nComponents}{Number of singular values to compute}

\item{binarize}{binarize the matrix}

\item{nFeatures}{Getting top N features for download analysis}

\item{...}{Arguments passed to seurat RunLSI}
}
\value{
Seurat object
}
\description{
LSI reduction analysis
}
