% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak.R
\name{summit2peak}
\alias{summit2peak}
\title{extent summit to build a peak set}
\usage{
summit2peak(summits, nSummits = 2e+05, extend = 250,
  chromSizes = NULL, blacklist = NULL)
}
\arguments{
\item{summits}{summit file generated from MACS2 callpeak}

\item{nSummits}{number of summit to use}

\item{extend}{extend the summit}

\item{chromSizes}{chromosome length}

\item{blacklist}{genome black list in BED format}
}
\description{
extent summit to build a peak set
}
