% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_peak_analysis.R
\name{uniqueFeatures}
\alias{uniqueFeatures}
\title{Find the unique feature in clusters}
\usage{
uniqueFeatures(mat, groups, minSdRatio = 0.001, minLFC = 0.25,
  zCutoff = 1, breakPt = "last", padj = 0.01, padjMethod = "fdr",
  clusterCols = TRUE, sparse = FALSE, twoWay = FALSE,
  groupMin = 10, minGroupSize = 1, maxGroupSize = NULL)
}
\arguments{
\item{mat}{matrix object}

\item{groups}{cluster groups}

\item{minSdRatio}{min sd}

\item{minLFC}{min logFoldChange}

\item{zCutoff}{z-score cutoff}

\item{breakPt}{break point}

\item{padj}{adjusted pvalue threshold}

\item{padjMethod}{pvalue adjust methods, see p.adjust.methods}

\item{clusterCols}{cluster the columns}

\item{sparse}{set it to TRUE is mat is sparse matrix}

\item{twoWay}{two way}

\item{groupMin}{grou min}

\item{minGroupSize}{min group size}

\item{maxGroupSize}{max group size}
}
\description{
Find the unique feature in clusters
}
