% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{writeFragmentsToBed}
\alias{writeFragmentsToBed}
\title{Write the fragments to BED}
\usage{
writeFragmentsToBed(gr, subset = NULL, filename = NULL,
  append = FALSE)
}
\arguments{
\item{gr}{GRanges which store the framgments postion information}

\item{subset}{subset the GRanges}

\item{filename}{output filename}

\item{append}{If TRUE, the file is opened in append mode and column names (header row) are not written.}
}
\description{
Write the fragments to BED
}
