% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak.R
\name{runMACS2}
\alias{runMACS2}
\title{Run MACS2}
\usage{
runMACS2(clusterResults, dirClusters = "results/LSI-Cluster-Beds",
  dirPeaks = "results/LSI-Cluster-Peaks", macs2Path = "macs2",
  method = "q", cutoff = 0.05, shift = -75, extsize = 150,
  bedGraph = TRUE, genomeSize = 2.7e+09, numCores = 6)
}
\arguments{
\item{macs2Path}{macs2 path}

\item{method}{q qvalue, p pvalue}

\item{cutoff}{cutoff of pvalue or qvalue}

\item{shift}{The arbitrary shift in bp}

\item{extsize}{The arbitrary extension size in bp}

\item{genomeSize}{Effective genome size}
}
\description{
Run MACS2
}
