% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFragmentsToBed.R
\name{writeFragmentsToBed}
\alias{writeFragmentsToBed}
\title{Write the fragments per cluster}
\usage{
writeFragmentsToBed(x, fragmentFiles = NULL,
  dirClusters = "results/LSI-Cluster-Beds/")
}
\arguments{
\item{x}{Seurat or SummarizedExperiment object}

\item{fragmentFiles}{vector store path of fragments GRanges Rds}

\item{dirClusters}{output directory}
}
\description{
Write the fragments per cluster
}
